requirements.txt
scrapy
boto3
elementary-data
python-dotenv
hashids
pyarrow
pandas
psycopg2-binary
dbt-postgres
polars[sqlalchemy]
duckdb
msgspec
connectorx
adbc-driver-postgresql
numpy
setuptools